Rename `defaultHost.ts` file & fix a wrong spelling
Rename `src/defaultHost.ts` to `src/constant.ts`
and fix the wrong spelling `defaulPort` typo by
just ranaming it to `defaultPort`.
