+++ src/constant.ts
@@ -1,0 +1,2 @@
+export const defaultPort = '11434';
+export const defaultHost = `http://127.0.0.1:${defaultPort}`;
