--- src/defaultHost.ts
@@ -1,2 +1,0 @@
-export const defaulPort = '11434';
-export const defaultHost = `http://127.0.0.1:${defaulPort}`;
